backend/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.9
# by the following command:
#
#    pip-compile
#
asgiref==3.7.1
    # via
    #   -r requirements.in
    #   django
    #   djoser
black==23.3.0
    # via -r requirements.in
build==0.10.0
    # via
    #   -r requirements.in
    #   pip-tools
certifi==2023.5.7
    # via
    #   -r requirements.in
    #   requests
cffi==1.15.1
    # via
    #   -r requirements.in
    #   cryptography
charset-normalizer==3.1.0
    # via
    #   -r requirements.in
    #   requests
click==8.1.3
    # via
    #   -r requirements.in
    #   black
    #   pip-tools
coreapi==2.3.3
    # via
    #   -r requirements.in
    #   djoser
coreschema==0.0.4
    # via
    #   -r requirements.in
    #   coreapi
cryptography==40.0.2
    # via
    #   -r requirements.in
    #   social-auth-core
defusedxml==0.7.1
    # via
    #   -r requirements.in
    #   python3-openid
    #   social-auth-core
django==3.2.9
    # via
    #   -r requirements.in
    #   django-filter
    #   djangorestframework
    #   djangorestframework-simplejwt
django-filter==23.2
    # via -r requirements.in
django-templated-mail==1.1.1
    # via
    #   -r requirements.in
    #   djoser
djangorestframework==3.14.0
    # via
    #   -r requirements.in
    #   djangorestframework-simplejwt
djangorestframework-simplejwt==4.7.2
    # via
    #   -r requirements.in
    #   djoser
djoser==2.1.0
    # via -r requirements.in
flake8==6.0.0
    # via
    #   -r requirements.in
    #   flake8-broken-line
    #   flake8-isort
    #   pep8-naming
flake8-broken-line==1.0.0
    # via -r requirements.in
flake8-isort==6.0.0
    # via -r requirements.in
flake8-plugin-utils==1.3.2
    # via
    #   -r requirements.in
    #   flake8-return
flake8-return==1.2.0
    # via -r requirements.in
gunicorn==20.1.0
    # via -r requirements.in
idna==3.4
    # via
    #   -r requirements.in
    #   requests
isort==5.12.0
    # via
    #   -r requirements.in
    #   flake8-isort
itypes==1.2.0
    # via
    #   -r requirements.in
    #   coreapi
jinja2==3.1.2
    # via
    #   -r requirements.in
    #   coreschema
markupsafe==2.1.2
    # via
    #   -r requirements.in
    #   jinja2
mccabe==0.7.0
    # via
    #   -r requirements.in
    #   flake8
mypy-extensions==1.0.0
    # via
    #   -r requirements.in
    #   black
oauthlib==3.2.2
    # via
    #   -r requirements.in
    #   requests-oauthlib
    #   social-auth-core
packaging==23.1
    # via
    #   -r requirements.in
    #   black
    #   build
pathspec==0.11.1
    # via
    #   -r requirements.in
    #   black
pep8-naming==0.13.3
    # via -r requirements.in
pillow==9.5.0
    # via
    #   -r requirements.in
    #   reportlab
pip-tools==6.13.0
    # via -r requirements.in
platformdirs==3.5.1
    # via
    #   -r requirements.in
    #   black
psycopg2-binary==2.9.6
    # via -r requirements.in
pycodestyle==2.10.0
    # via
    #   -r requirements.in
    #   flake8
pycparser==2.21
    # via
    #   -r requirements.in
    #   cffi
pyflakes==3.0.1
    # via
    #   -r requirements.in
    #   flake8
pyjwt==2.7.0
    # via
    #   -r requirements.in
    #   djangorestframework-simplejwt
    #   social-auth-core
pyproject-hooks==1.0.0
    # via
    #   -r requirements.in
    #   build
python3-openid==3.2.0
    # via
    #   -r requirements.in
    #   social-auth-core
pytz==2023.3
    # via
    #   -r requirements.in
    #   django
    #   djangorestframework
reportlab==4.0.4
    # via -r requirements.in
requests==2.31.0
    # via
    #   -r requirements.in
    #   coreapi
    #   requests-oauthlib
    #   social-auth-core
requests-oauthlib==1.3.1
    # via
    #   -r requirements.in
    #   social-auth-core
six==1.16.0
    # via
    #   -r requirements.in
    #   social-auth-app-django
social-auth-app-django==4.0.0
    # via
    #   -r requirements.in
    #   djoser
social-auth-core==4.4.2
    # via
    #   -r requirements.in
    #   social-auth-app-django
sqlparse==0.4.4
    # via
    #   -r requirements.in
    #   django
tomli==2.0.1
    # via
    #   -r requirements.in
    #   black
    #   build
    #   pyproject-hooks
typing-extensions==4.6.2
    # via
    #   -r requirements.in
    #   asgiref
    #   black
uritemplate==4.1.1
    # via
    #   -r requirements.in
    #   coreapi
urllib3==2.0.2
    # via
    #   -r requirements.in
    #   requests
v==1
    # via -r requirements.in
wheel==0.40.0
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
